// src/api/generated/model/importCompaniesCsvMapping.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type ImportCompaniesCsvMapping = typeof ImportCompaniesCsvMapping[keyof typeof ImportCompaniesCsvMapping];


export const ImportCompaniesCsvMapping = {
  default: 'default',
  crunchbase: 'crunchbase',
  custom: 'custom',
} as const;